feat: Made seperate compoents from SidePanel
